chore: remove stub for api proxy
