+++ clients/api-proxy/greffe/index.ts
@@ -1,5 +1,4 @@
 import routes from '#clients/routes';
-import stubClientWithSnapshots from '#clients/stub-client-with-snaphots';
 import constants from '#models/constants';
 import {
   IUniteLegale,
@@ -9,39 +8,17 @@ import {
 import { Siren, formatDateYear } from '#utils/helpers';
 import { etatFromEtatAdministratifInsee } from '#utils/helpers/insee-variables';
 import { clientAPIProxy } from '../client';
-
-type IIGResponse = {
-  siren: Siren;
-  nomComplet: string;
-  etat: string;
-  natureJuridique: string;
-  activitePrincipale: string;
-  libelleActivitePrincipale: string;
-  dateCreation: string;
-  siege: string;
-  association: {
-    idAssociation: string | null;
-  };
-  immatriculation: {
-    dateDebutActivite: string;
-    dateFin: string;
-    duree: 0;
-    natureEntreprise: string[];
-    dateCloture: string;
-    dateImmatriculation: string;
-    dateRadiation: string;
-    isPersonneMorale: boolean;
-    capital: string;
-  };
-};
+import { IIGResponse } from './types';
 
 /**
  * Call EORI to validate a French EORI number
  * @param siret
  */
-const clientUniteLegaleIG = async (siren: Siren): Promise<IUniteLegale> => {
+export const clientUniteLegaleIG = async (
+  siren: Siren
+): Promise<IUniteLegale> => {
   return mapToDomainObject(
-    await clientAPIProxy<IIGResponse>(routes.proxy.ig + siren, {
+    await clientAPIProxy<IIGResponse>(routes.proxy.ig(siren), {
       timeout: constants.timeout.XL,
     }),
     siren
@@ -58,9 +35,3 @@ const mapToDomainObject = (r: IIGResponse, siren: Siren): IUniteLegale => {
     etatAdministratif: etatFromEtatAdministratifInsee(r.etat, r.siren),
   };
 };
-
-const stubbedClient = stubClientWithSnapshots({
-  clientUniteLegaleIG,
-});
-
-export { stubbedClient as clientUniteLegaleIG };
