+++ clients/api-proxy/rne/documents.ts
@@ -1,5 +1,4 @@
 import routes from '#clients/routes';
-import stubClientWithSnapshots from '#clients/stub-client-with-snaphots';
 import constants from '#models/constants';
 import { IActesRNE } from '#models/rne/types';
 import { Siren } from '#utils/helpers';
@@ -11,7 +10,7 @@ import { clientAPIProxy } from '../client';
  * RNE through the API proxy
  * @param siren
  */
-const fetchDocumentsFromRNE = async (siren: Siren) => {
+export const clientDocuments = async (siren: Siren) => {
   const route = routes.proxy.rne.documents.list + siren;
 
   const callerInfos = await sensitiveRequestCallerInfos();
@@ -22,24 +21,14 @@ const fetchDocumentsFromRNE = async (siren: Siren) => {
   });
 };
 
-const clientDownloadActe = async (id: string) =>
+export const clientDownloadActe = async (id: string) =>
   clientAPIProxy<string>(routes.proxy.rne.documents.download.acte + id, {
     timeout: constants.timeout.XXXXL,
     responseType: 'arraybuffer',
   });
 
-const clientDownloadBilan = async (id: string) =>
+export const clientDownloadBilan = async (id: string) =>
   clientAPIProxy<string>(routes.proxy.rne.documents.download.bilan + id, {
     timeout: constants.timeout.XXXXL,
     responseType: 'arraybuffer',
   });
-
-const stubbedClient = stubClientWithSnapshots({
-  fetchDocumentsFromRNE,
-});
-
-export {
-  stubbedClient as clientDocuments,
-  clientDownloadActe,
-  clientDownloadBilan,
-};
