+++ cypress/mocks/handlers/association.ts
@@ -1,0 +1,5 @@
+import { HttpResponse, HttpResponseResolver } from 'msw';
+
+export const associationHandler: HttpResponseResolver = ({ request }) => {
+  return HttpResponse.json({});
+};
