+++ cypress/mocks/handlers/ig.ts
@@ -1,0 +1,5 @@
+import { HttpResponse, HttpResponseResolver } from 'msw';
+
+export const igHandler: HttpResponseResolver = ({ request }) => {
+  return HttpResponse.json({});
+};
