+++ cypress/mocks/handlers/rne.ts
@@ -1,0 +1,9 @@
+import { HttpResponse, HttpResponseResolver } from 'msw';
+
+export const rneDefaultHandler: HttpResponseResolver = ({ request }) => {
+  return HttpResponse.json({});
+};
+
+export const rneFallbackHandler: HttpResponseResolver = ({ request }) => {
+  return HttpResponse.json({});
+};
